Fix concurrency, exception-safety, and correctness bugs in JNI layer and Java API
JNI / native fixes:
- Guard rerank / embedding / streaming to_json() calls against uncaught C++
  exceptions crossing the JNI boundary, which would otherwise abort the JVM.
- Hold a jllama_context user reference for the attached NativeServer worker so
  closing the model cannot free the context out from under the running HTTP
  frontend (use-after-free).
- Fix close()-during-inference hang: add a closing flag that unblocks readers
  parked in next()/wait_for_all(), and decrement the user count under the
  context mutex so the teardown wait cannot free the mutex/condvar before the
  releaser notifies (latent use-after-free in the ref-count itself).
- Fix NativeServer jstring_to_utf8 calling getBytes with a Charset argument
  while resolving the String overload (deterministic break on every server
  start); cache the JNI IDs thread-safely.
- Initialize parse_string_array slots to nullptr so an OOM path cannot free
  uninitialized memory.
- Add a pending-exception check in parse_jstring.

Java fixes:
- Widen token counts to long in the protocol translators and
  ChatResponseParser.extractUsageField to avoid truncation above ~2.1B tokens.
- Honor a pre-cancelled CancellationToken in complete() (stop resetting it at
  entry); fix completeAsJson javadoc (the params object is not mutated).

Tests and docs:
- Extract the OuteTTS codec filter into a testable helper; add rerank bounds
  and codec-filter unit tests, and a model-gated close()-during-inference test.
- Remove the dead BUILD_SHARED_LIBS ON line in CMakeLists.txt.
- Update plan and TODO docs to reflect the real teardown design.

+++ llama/src/main/java/net/ladenthin/llama/server/AnthropicApiSupport.java
@@ -254,14 +254,14 @@ final class AnthropicApiSupport {
             stopReason = anthropicStopReason(choice.path("finish_reason").asText("stop"));
             JsonNode openAiUsage = completion.path("usage");
             if (openAiUsage.isObject()) {
-                int promptTokens = openAiUsage.path("prompt_tokens").asInt(0);
-                int cachedTokens = openAiUsage
+                long promptTokens = openAiUsage.path("prompt_tokens").asLong(0);
+                long cachedTokens = openAiUsage
                         .path("prompt_tokens_details")
                         .path("cached_tokens")
-                        .asInt(0);
-                usage.put("input_tokens", Math.max(0, promptTokens - cachedTokens));
+                        .asLong(0);
+                usage.put("input_tokens", Math.max(0L, promptTokens - cachedTokens));
                 usage.put("cache_read_input_tokens", cachedTokens);
-                usage.put("output_tokens", openAiUsage.path("completion_tokens").asInt(0));
+                usage.put("output_tokens", openAiUsage.path("completion_tokens").asLong(0));
             }
         } catch (IOException e) {
             stopReason = "end_turn";
@@ -401,7 +401,7 @@ final class AnthropicApiSupport {
     }
 
     /** Final message delta carrying token usage collected from the trailing OpenAI usage chunk. */
-    static String messageDeltaEvent(String stopReason, int inputTokens, int outputTokens, int cachedTokens) {
+    static String messageDeltaEvent(String stopReason, long inputTokens, long outputTokens, long cachedTokens) {
         ObjectNode data = OBJECT_MAPPER.createObjectNode();
         data.put("type", "message_delta");
         ObjectNode delta = data.putObject("delta");
